sample ua_detrac video added
