+++ test/integration_tests/test_select_executor.py
@@ -57,3 +57,32 @@ class SelectExecutorTest(unittest.TestCase):
         actual_batch = [perform_query(select_query)]
         expected_batch = list(create_dummy_batches())
         self.assertTrue(actual_batch, expected_batch)
+
+    def test_select_and_where_video_in_table(self):
+        load_query = """LOAD DATA INFILE 'dummy.avi' INTO MyVideo;"""
+        perform_query(load_query)
+        select_query = "SELECT id,data FROM MyVideo WHERE id = 5;"
+        actual_batch = perform_query(select_query)
+        expected_batch = list(create_dummy_batches(filters=[5]))[0]
+        self.assertTrue(actual_batch, expected_batch)
+
+        select_query = "SELECT data FROM MyVideo WHERE id = 5;"
+        actual_batch = perform_query(select_query)
+        expected_rows = [{"data": np.array(
+            np.ones((2, 2, 3)) * 0.1 * float(5 + 1) * 255, dtype=np.uint8)}]
+        expected_batch = Batch(frames=pd.DataFrame(expected_rows))
+        self.assertTrue(actual_batch, expected_batch)
+
+        select_query = "SELECT id, data FROM MyVideo WHERE id >= 2;"
+        actual_batch = perform_query(select_query)
+        expected_batch = list(
+            create_dummy_batches(
+                filters=range(
+                    2, NUM_FRAMES)))[0]
+        self.assertTrue(actual_batch, expected_batch)
+
+        select_query = "SELECT id, data FROM MyVideo WHERE id >= 2 AND id < 5;"
+        actual_batch = perform_query(select_query)
+        expected_batch = list(create_dummy_batches(filters=range(2, 5)))[0]
+
+        self.assertTrue(actual_batch, expected_batch)
