--- test/integration_tests/test_where_executor.py
@@ -1,66 +1,0 @@
-# coding=utf-8
-# Copyright 2018-2020 EVA
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-import unittest
-import os
-import numpy as np
-import pandas as pd
-
-from src.catalog.catalog_manager import CatalogManager
-from src.models.storage.batch import Batch
-
-from test.util import create_sample_video
-from test.util import create_dummy_batches, perform_query
-
-NUM_FRAMES = 10
-
-
-class WhereExecutorTest(unittest.TestCase):
-
-    def setUp(self):
-        CatalogManager().reset()
-        create_sample_video()
-
-    def tearDown(self):
-        os.remove('dummy.avi')
-
-    # integration test
-    def test_select_and_where_video_in_table(self):
-        load_query = """LOAD DATA INFILE 'dummy.avi' INTO MyVideo;"""
-        perform_query(load_query)
-        select_query = "SELECT id,data FROM MyVideo WHERE id = 5;"
-        actual_batch = perform_query(select_query)
-        expected_batch = list(create_dummy_batches(filters=[5]))[0]
-        self.assertTrue(actual_batch, expected_batch)
-
-        select_query = "SELECT data FROM MyVideo WHERE id = 5;"
-        actual_batch = perform_query(select_query)
-        expected_rows = [{"data": np.array(
-            np.ones((2, 2, 3)) * 0.1 * float(5 + 1) * 255, dtype=np.uint8)}]
-        expected_batch = Batch(frames=pd.DataFrame(expected_rows))
-        self.assertTrue(actual_batch, expected_batch)
-
-        select_query = "SELECT id, data FROM MyVideo WHERE id >= 2;"
-        actual_batch = perform_query(select_query)
-        expected_batch = list(
-            create_dummy_batches(
-                filters=range(
-                    2, NUM_FRAMES)))[0]
-        self.assertTrue(actual_batch, expected_batch)
-
-        select_query = "SELECT id, data FROM MyVideo WHERE id >= 2 AND id < 5;"
-        actual_batch = perform_query(select_query)
-        expected_batch = list(create_dummy_batches(filters=range(2, 5)))[0]
-
-        self.assertTrue(actual_batch, expected_batch)
